Implement backend detection model and drawing predictions
Quite a few things to talk about in this commit, but I should
keep it brief. The comments, and code are a bit of a mess at the
moment. Will need to clean that up in a later commit.

First, the backend is successfully able to load a PyTorch
pre-trained object detection model and returns predictions on
an image. I'm unsure why, but it isn't filtering those predictions
based on confidence, so there's still dozens of low confidence
predictions that are basically just noise being returned.

Additionally, I created another temporary api endpoint that was
successfully able to take a user inputted image, draw bounding
boxes on it, and then return and rerender it on the React frontend.
It ended up being fairly complicated, and will likely need to be
tested more rigorously, but for now it's a working solution.